Ability to handle assertion definitions.
Assertion definitions are methods that can be called directly on a page
or activity definition. These methods assert some aspect of the page or
activity. Once the asserted aspect is verified, the asserted aspect is
stored as a variable on the definition class.

It is important to understand that the asserted state is defined on the
page or activity definition class, while any use of that state is
handled on an instance of the page or activity definition.