Ability to handle element definitions.
Allow a page definition to assert an element definition. The element
definition represents an actual web object on a given page. The
friendly names given to these definitions serve as identifiers that
are proxied to the browser driver. This allows the methods of the
browser driver to be called directly on the element definition.