Removal of multi-platform approach.
Symbiont will only support one driver library, which will be the
Watir-WebDriver library. Multi-platform support simply did not make
much sense given that Watir-WebDriver is a wrapper around Selenium.
This means it will be possible to drop down to Selenium if and when
necessary. Further, abstracting away the specific drivers in order
to provide a common interface seemed to take away from using the
specific strengths of particular a given driver all in the name of
providing a common interface. The maintenance headaches and code
bloat of supporting multiple platforms mitigated against the approach.