Watir container module used to generate element methods.
Since Symbiont is focusing on Watir-WebDriver as its sole driver library,
this allows for calling Watir directly to determine what web elements it
supports. That list of elements is then used as the basis for defining a
series of methods that correspond to those element names. Those methods
can then be used to construct element definitions on page classes.