Driver-based actions available to elements.
Allow instanced element definitions to be able to access many of the
methods that are accessible directly on a Watir::Browser driver instance.
This allows for those same methods to be available via an instance of an
element definition rather than having to call the driver specifically.
This keeps the context of the most relevant actions relative to the page
definition, which encapsulates the driver.