Data setter mechanism added.
The data setter module provides a convenience mechanism for allowing data
conditions to be specified and passed in to a page or activity definition.
Text fields, select lists, checkboxes and radios can be populated based on
the data indicated.

The data setter is the beginning stage of a more comprehensive data builder
mechanism.