Data builder mechanism added.
The data builder module provides a convenience mechanism that builds off of
the data setter module. The data builder is used to build data from common
repositories of test data, which will be stored in yaml files. Symbiont will
look for a default data file but you can specify your own as well.